requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
groq==0.25.0
python-dotenv==1.0.1
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
langchain-openai==0.0.2
pillow==10.0.0
beautifulsoup4==4.12.2 